ADA-102 Adding meeting attendance to exec dashboard
(cherry picked from commit ff4368722f2cd0abef0a07760b4fef080fa21c92)
