+++ acmw-db-app/styles/Database.module.css
@@ -30,3 +30,11 @@
   justify-content: left;
   align-items: left;
 }
+
+.table {
+  border: 2px solid #373738;
+  padding: 10px;
+  margin: 10px 0px;
+  background-color: white;
+  color: #373738;
+}
